Wire up insert / select

+++ src/query/ast.rs
@@ -1,4 +1,4 @@
 pub enum Query<'a> {
     Select,
-    Insert(u64, &'a str, &'a str),
+    Insert(u32, &'a str, &'a str),
 }
